refactor: extract PostTranscriptionFailurePresenter.swift from AppErrorSubPresenters.swift
Splits `PostTranscriptionFailurePresenter` @MainActor class (~24
lines) out of AppErrorSubPresenters.swift into its own file so that
file holds only `TranscriptPersistenceFailurePresenter`.

Byte-identical move of lines 42-65. Both files import Foundation.

AppErrorSubPresenters.swift: 65 → 40 lines. Tests: 667.
