+++ Sources/BugNarrator/Services/AppErrorSubPresenters.swift
@@ -38,28 +38,3 @@ final class TranscriptPersistenceFailurePresenter {
         showTranscriptWindow()
     }
 }
-
-@MainActor
-final class PostTranscriptionFailurePresenter {
-    private let errorPresenter: AppErrorPresenter
-    private let showSettingsWindow: () -> Void
-
-    init(
-        errorPresenter: AppErrorPresenter,
-        showSettingsWindow: @escaping () -> Void
-    ) {
-        self.errorPresenter = errorPresenter
-        self.showSettingsWindow = showSettingsWindow
-    }
-
-    func present(
-        _ error: Error,
-        operation: AppErrorOperation = .postTranscription
-    ) {
-        let result = errorPresenter.presentPostTranscriptionError(error, operation: operation)
-        if result.shouldOpenSettingsWindow {
-            showSettingsWindow()
-        }
-    }
-}
-
